nb-omnibus-router/requirements.txt
# NeuralBlitz Omnibus Router Dependencies
# Generated: 2026-02-08

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Async Support
aiofiles==23.2.1
httpx==0.26.0

# Configuration
pyyaml==6.0.1

# Logging
loguru==0.7.2

# CLI
typer==0.9.0
rich==13.7.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
